+++ infra/src/main/kotlin/org/yapp/infra/InfraBaseConfigGroup.kt
@@ -1,6 +1,6 @@
 package org.yapp.infra
 
-import org.yapp.infra.config.external.oauth.RestClientConfig
+import org.yapp.infra.config.external.api.RestClientConfig
 import org.yapp.infra.config.external.redis.RedisConfig
 import org.yapp.infra.config.internal.async.AsyncConfig
 import org.yapp.infra.config.internal.jpa.JpaConfig
@@ -11,5 +11,5 @@ enum class InfraBaseConfigGroup(
     ASYNC(AsyncConfig::class.java),
     JPA(JpaConfig::class.java),
     REDIS(RedisConfig::class.java),
-    OAUTH(RestClientConfig::class.java)
+    REST_CLIENT(RestClientConfig::class.java)
 }
