+++ infra/src/main/kotlin/org/yapp/infra/config/external/api/RestClientConfig.kt
@@ -1,4 +1,4 @@
-package org.yapp.infra.config.external.oauth
+package org.yapp.infra.config.external.api
 
 import org.apache.hc.client5.http.config.RequestConfig
 import org.apache.hc.client5.http.impl.classic.CloseableHttpClient
@@ -6,15 +6,29 @@ import org.apache.hc.client5.http.impl.classic.HttpClients
 import org.apache.hc.core5.util.Timeout
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Primary
 import org.springframework.http.client.HttpComponentsClientHttpRequestFactory
 import org.springframework.web.client.RestClient
 import org.yapp.infra.InfraBaseConfig
 
 @Configuration
+
 class RestClientConfig : InfraBaseConfig {
 
     @Bean
-    fun restClient(): RestClient {
+    @Primary
+    fun generalRestClient(): RestClient {
+        return createConfiguredRestClientBuilder().build()
+    }
+
+    @Bean("aladinApiRestClient")
+    fun aladinRestClient(): RestClient {
+        return createConfiguredRestClientBuilder()
+            .baseUrl("http://www.aladin.co.kr/ttb/api")
+            .build()
+    }
+
+    private fun createConfiguredRestClientBuilder(): RestClient.Builder {
         val requestConfig = RequestConfig.custom()
             .setConnectionRequestTimeout(Timeout.ofDays(5000))
             .setResponseTimeout(Timeout.ofDays(5000))
@@ -28,6 +42,5 @@ class RestClientConfig : InfraBaseConfig {
 
         return RestClient.builder()
             .requestFactory(factory)
-            .build()
     }
 }
